requirements.txt
requests==2.19.1
Flask==1.0.2
flask_cors==3.0.6
python-dotenv==0.9.1

